+++ application/app/Models/CatalystProfile.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Models;
+
+class CatalystProfile extends Model
+{
+    public $guarded = [];
+}